refactor(redux): migrate app settings dialog
