+++ src/store/index.js
@@ -19,7 +19,7 @@ import {
   createFilter,
 } from 'redux-persist-transform-filter';
 
-import { showAppSettingsDialog } from '~/actions/app-settings';
+import { showAppSettingsDialog } from '~/features/settings/actions';
 import { updateAveragingByIds } from '~/features/measurement/slice';
 import { shouldPreventSleepMode } from '~/features/power-saving/selectors';
 import { updateRTKStatistics } from '~/features/rtk/slice';
